Changed defaul min_posts_suggest to "1"
This is just a minor migration change, wich does nothing bad to existing extensions. It just sets the default value to 1, instead of 0, so the load on boards is reduced (null posters will not be mentionable).

I don't want to add this change to a new migration, this would overwrite the value in all installations, and this is still in development, so I think it is okay to do this little adjustment.